utils: simplify the code
